+++ src/Data/Curve.cpp
@@ -123,7 +123,7 @@ float Curve::get(int pos)
 	for (int i=1; i<points.num; i++)
 		if (pos < points[i].pos){
 			float dv = points[i].value - points[i-1].value;
-			float dp = points[i].pos - points[i-1].pos;
+			float dp = (float)(points[i].pos - points[i-1].pos);
 			return points[i-1].value + dv * (pos - points[i-1].pos) / dp;
 		}
 	return points.back().value;
